add variables symboliques and constantes symboliques

+++ td3/src/main/java/td3/ConstanteRationnelle.java
@@ -19,7 +19,7 @@ public final class ConstanteRationnelle implements ExpressionArithmetique {
 
 	@Override
 	public double calculer() {
-		return this.numerateur / this.denominateur;
+		return (double)this.numerateur / this.denominateur;
 	}
 
 	@Override
